Sample time parsing in MainSimulationTab fixed.

+++ DAMOS/trunk/plugins/org.eclipselabs.damos.simulation.ide.ui/src/org/eclipselabs/damos/simulation/ide/ui/internal/launch/MainSimulationTab.java
@@ -309,7 +309,7 @@ public class MainSimulationTab extends AbstractLaunchConfigurationTab {
 		try {
 			String sampleTime = sampleTimeText.getText();
 			if (!SimulationLaunchConfigurationDelegate.DEFAULT_SAMPLE_TIME.equals(sampleTime)) {
-				Long.parseLong(sampleTime);
+				Double.parseDouble(sampleTime);
 				configuration.setAttribute(SimulationLaunchConfigurationDelegate.ATTRIBUTE__SAMPLE_TIME, sampleTime);
 			}
 		} catch (NumberFormatException e) {
